requirements.txt
# Web framework
streamlit==1.39.0

# Data manipulation & math
pandas==2.2.3
numpy==2.1.2

# ML and model handling
scikit-learn==1.5.2
xgboost==3.0.5
imbalanced-learn==0.12.3

# Explainability (compatible with above)
shap
matplotlib==3.9.2

# Azure Blob access
azure-storage-blob==12.19.1

# Environment management
python-dotenv==1.0.1

# OpenAI SDK (latest, stable)
openai==1.74.0

xlsxwriter==3.2.9